refactor(component): 필터링 관련 공통 컴포넌트 폴더 분리
- @/app/_components/filter 폴더에 저장
- SizeToggleGroup 공통 컴포넌트로 리팩토링
